+++ src/pyday/__init__.py
@@ -1,4 +1,6 @@
-from .ChangLanguage import *
 from .DataReader.DataReader import *
 from .Data2PDF.Data2PDF import *
 from .DataVis.DataVis import *
+
+from .Tool.dirTree import *
+from .Tool.ChangLanguage import *