+++ src/test.py
@@ -1,3 +1,9 @@
+# from pyday import DataReader
+# dr = DataReader()
+# dr.inFile("train.csv")
+# dr.toFile("test.all")
+
 from pyday import Data2PDF
-d2pdf = Data2PDF("test2.json")
+d2pdf = Data2PDF()
+d2pdf.inFile("test2.json")
 d2pdf.toFile("test2.pdf")